feat(navigation): Added segments for screens & added initial route logic
