+++ src/screens/config/index.ts
@@ -1,1 +1,2 @@
+export * from './theme';
 export * from './types';
