+++ src/screens/model/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useInitialRoute } from './useInitialRoute';
