+++ src/screens/model/hooks/useInitialRoute.ts
@@ -1,0 +1,11 @@
+import { SessionModel } from '@entities/session';
+
+import { ScreenRoute } from '../../config';
+
+function useInitialRoute(): ScreenRoute {
+  const hasSession = SessionModel.useHasSession();
+
+  return hasSession ? 'Applets' : 'Login';
+}
+
+export default useInitialRoute;
