+++ src/screens/ui/AppletDetailsScreen.tsx
