+++ src/screens/ui/index.ts
@@ -1,0 +1,9 @@
+export { default as AppletsScreen } from './AppletsScreen';
+export { default as ChangeLanguageScreen } from './ChangeLanguageScreen';
+export { default as ChangePasswordScreen } from './ChangePasswordScreen';
+export { default as ForgotPasswordScreen } from './ForgotPasswordScreen';
+export { default as LoginScreen } from './LoginScreen';
+export { default as SignUpScreen } from './SignUpScreen';
+export { default as AboutScreen } from './AboutScreen';
+export { default as SettingsScreen } from './SettingsScreen';
+export { default as ActivityListScreen } from './ActivityListScreen';
