+++ src/shared/lib/index.ts
@@ -1,6 +1,5 @@
 export * from './constants';
 export * from './hooks';
-export * from './navigation';
 export * from './utils';
 export * from './storages';
 export * from './types';
